consolidated ingestion_config
